requirements.txt
pymupdf>=1.23.0
pdfplumber>=0.10.0
sentence-transformers>=2.7.0
transformers>=4.41.0,<5.0
faiss-cpu>=1.8.0
nltk>=3.8.0
streamlit>=1.35.0
pytest>=8.0.0
pytest-cov>=5.0.0
openai>=1.30.0
python-dotenv>=1.0.0
